requirements.txt
antlr4-python3-runtime==4.9.3
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==2.0.4
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
easydict==1.12
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.49.0
fqdn==1.5.1
fsspec==2024.2.0
ftfy==6.1.1
gensim==4.3.2
gitdb==4.0.11
GitPython==3.1.42
grpcio==1.62.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.10
ipdb==0.13.13
ipykernel==6.29.2
ipython==8.22.1
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lightning-utilities==0.10.1
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
mkl-fft==1.3.8
mkl-random==1.2.4
mkl-service==2.4.0
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
nltk==3.8.1
notebook==7.0.8
notebook_shim==0.2.4
numpy==1.26.3
omegaconf==2.3.0
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pillow==10.2.0
pip==23.3.1
platformdirs==4.2.0
prettytable==3.10.0
prometheus_client==0.20.0
promise==2.3
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rpds-py==0.18.0
scipy==1.12.0
Send2Trash==1.8.2
sentry-sdk==1.40.5
setproctitle==1.3.3
setuptools==68.2.2
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
terminado==0.18.0
textaugment==2.0.0
textblob==0.18.0.post0
timm==0.6.11
tinycss2==1.2.1
tomli==2.0.1
torchmetrics==1.3.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==2.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
wheel==0.41.2
widgetsnbextension==4.0.10
